Renamed mongo_helper.py to connection.py
